Moved star rating to bookfinds/widgets
